requirements.txt
anthropic==0.86.0
PyGithub==2.9.0
pydantic==2.12.5